perf: shrink Position to u32 offset; Result<Option<Event>, Error> now 12 bytes
Headline: bourne now matches or beats serde_json on 7 of 10 workloads,
including `typed_struct`, `stream ints/10000`, `vec_borrowed_str/10000`,
`vec_string/10000`. Vs prior baseline:

  workload                     change   vs serde_json
  stream small_object          -58%     bourne 360ns / serde 661ns  (1.84x faster)
  stream ints/10000            -60%     bourne 206us / serde 211us  (parity)
  stream strings/10000         -67%     bourne 269us / serde 595us  (2.21x faster)
  typed_struct                 -63%     bourne 279ns / serde 278ns  (parity)
  vec_i64/10000                -75%     bourne 178us / serde 130us  (1.36x slower)
  vec_borrowed_str/10000       -65%     bourne 205us / serde 242us  (1.18x faster)
  vec_string/10000             -47%     bourne 537us / serde 548us  (parity)

Originally vec_i64/10000 was 5.4x slower; now 1.36x.

The change
==========

`Position` was `(offset: usize, line: u32, column: u32)` = 16 bytes.
That made `Error = (ErrorKind, Position)` = 24 bytes, which made
`Result<Option<Event>, Error>` = 24 bytes. 24 bytes won't return in two
GP registers under the System V ABI, so every `next_event()` return
spilled through stack memory. Profile evidence: ~14% of the typed
integer-array hot loop was these stack moves.

Now `Position` is just `{ offset: u32 }` (4 bytes). `Error` is 8 bytes.
`Result<Option<Event>, Error>` is 12 bytes — fits in two registers.
Line and column are reconstructed on demand via:

    pub fn resolve(&self, input: &[u8]) -> LineColumn { ... }

The scan only fires when an error is actually rendered or the consumer
explicitly asks for line/column — never inside the parser hot loop.

Knock-on changes:
* `Position` exposes only `offset: u32`. Callers that read `line`/
  `column` directly now call `position.resolve(input).line`.
* `Position::START` has no line/column. The Display impl shows just
  the byte offset.
* All call sites updated; property test fixed up to compare `offset
  as usize` to `bytes.len()`.

+++ crates/bourne/tests/properties.rs
@@ -62,7 +62,7 @@ proptest! {
                 Ok(None) => break,
                 Err(e) => {
                     prop_assert!(
-                        e.position.offset <= bytes.len(),
+                        (e.position.offset as usize) <= bytes.len(),
                         "error position {} out of bounds for input of {} bytes",
                         e.position.offset, bytes.len(),
                     );
